refactor(ui): consolidate timestamp helpers
