+++ apps/ui/src/app/views/esp_flash_feature_presenter.ts
@@ -3,8 +3,8 @@ import type {
   EspFlashStatusPayload,
   EspSerialPortPayload,
 } from "../../transport/http_models";
+import { formatEpochTimestamp } from "../../format";
 import type { VisualVariant } from "../view_style_types";
-import { formatEpochTimestamp } from "./dom_helpers";
 import type { MaintenanceReadinessPanelModel } from "./maintenance_readiness_view";
 import type {
   EspFlashHistoryAttemptModel,
